sudo to target process UID
When the user of the PID to attach to is not a sudoer, one would expect
to be able to run the scripts as root. However due to the JDK bug
https://bugs.openjdk.java.net/browse/JDK-8036559, root cannot attach to
the JVM of another user. To address this, I’ve added a sudo to the
target user, so
the scripts can be run as root. This should be a no-op if you are
already that user.